fix(global): fix cypress export
