+++ .sonarcloud.properties
@@ -1,1 +1,1 @@
-sonar.exclusions=config/cypress/**
+sonar.exclusions=config/cypress/**,src/cypress/**
