+++ package.json
@@ -86,6 +86,7 @@
     "@typescript-eslint/eslint-plugin": "6.21.0",
     "@typescript-eslint/parser": "8.35.1",
     "@vitejs/plugin-react": "^5.0.0",
+    "@vitejs/plugin-react-swc": "^4.1.0",
     "browser-assert": "1.2.1",
     "chromatic": "11.28.2",
     "classnames": "2.5.1",
