+++ src/cypress/commands/clickButton.ts
