+++ src/cypress/commands/clickLink.ts
