+++ src/cypress/commands/clickOutside.ts
