+++ src/cypress/commands/fill/checkCheckbox.ts
