+++ src/cypress/commands/fill/checkMultiCheckboxOptions.ts
