+++ src/cypress/commands/forceCheck.ts
