+++ src/cypress/commands/forceClear.ts
