+++ src/cypress/commands/forceClick.ts
