+++ src/cypress/commands/forceUncheck.ts
