+++ src/cypress/commands/getTableRowById.ts
