+++ src/cypress/commands/getTableRowByText.ts
