+++ src/cypress/tsconfig.dist.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["**/*.ts"],
+  "exclude": ["**/*.test.ts"]
+}
