+++ src/cypress/utils/ensureElement.ts
