+++ src/cypress/utils/findElementBySelector.ts
