+++ src/cypress/utils/findElementByText.ts
