+++ src/cypress/utils/throwError.ts
